# docs/swagger.yaml
basePath: /
definitions:
...
  service.DeleteComment:
    properties:
      commentid:
        type: integer
      topicid:
        type: integer
    type: object
  service.DeleteReply:
    properties:
      replyid:
        type: integer
      userid:
        type: string
    type: object
  service.DeleteTopic:
    properties:
      sectionid:
        type: integer
      topic:
        type: integer
      userid:
        type: string
    type: object
...
  service.Reply:
    properties:
      commentid:
        type: integer
      content:
        type: string
      pid:
        type: integer
      userid:
        type: string
    type: object
  service.Topic:
    properties:
      content:
        type: string
      evaluateStatus:
        type: integer
      favoriteStatus:
        type: integer
      istop:
        type: integer
      sectionid:
        type: integer
      title:
        type: string
      userid:
        type: string
    type: object
  service.UpdateComment:
    properties:
      commentid:
        type: integer
      content:
        type: string
    type: object
  service.UpdateReply:
    properties:
      content:
        type: string
      replyid:
        type: integer
      userid:
        type: string
    type: object
  service.UpdateTopic:
    properties:
      content:
        type: string
      istop:
        type: integer
      sectionid:
        type: integer
      title:
        type: string
      topic:
        type: integer
      userid:
        type: string
    type: object
...
  utils.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  utils.Message:
    properties:
      code:
        example: 1
        type: integer
      data:
        type: object
      message:
        example: ok
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: revolveyao@qq.com
    name: API支持者
    url: http://www.swagger.io/support
  description: 星辰大海后端API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 星辰大海 API
  version: "1.0"
paths:
...
  /v1/comment/delete:
    post:
      consumes:
      - application/json
      description: 删除评论
      operationId: get-topic-content-json
      parameters:
      - description: 删除评论传入json
        in: body
        name: topic
        required: true
        schema:
          $ref: '#/definitions/service.DeleteComment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 删除评论
...
  /v1/comment/update:
    post:
      consumes:
      - application/json
      description: 获取评论内容
      operationId: get-topic-content-json
      parameters:
      - description: 更新评论传入json
        in: body
        name: topic
        required: true
        schema:
          $ref: '#/definitions/service.UpdateComment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 更新评论
...
  /v1/reply/add:
    post:
      consumes:
      - application/json
      description: 添加回复
      operationId: post-reply-add-json
      parameters:
      - description: 回复
        in: body
        name: reply
        required: true
        schema:
          $ref: '#/definitions/service.Reply'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 添加回复
  /v1/reply/delete:
    post:
      consumes:
      - application/json
      description: 删除回复
      operationId: get-topic-content-json
      parameters:
      - description: 删除回复传入json
        in: body
        name: topic
        required: true
        schema:
          $ref: '#/definitions/service.DeleteReply'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 删除回复
  /v1/reply/get:
    get:
      consumes:
      - application/json
      description: 根据页码等数据获取指定评论的回复
      operationId: get-reply-page-json
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pagesize
        type: integer
      - description: 总数
        in: query
        name: total
        type: integer
      - description: 排序方式
        in: query
        name: sort
        type: string
      - description: 评论ID
        in: query
        name: commentid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: 获取回复
  /v1/reply/get/{userid}:
    get:
      consumes:
      - application/json
      description: 根据页码等数据获取指定评论的回复-带有用户权限
      operationId: get-reply-page-user-json
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pagesize
        type: integer
      - description: 总数
        in: query
        name: total
        type: integer
      - description: 排序方式
        in: query
        name: sort
        type: string
      - description: 评论ID
        in: query
        name: commentid
        required: true
        type: integer
      - description: 用户ID
        in: path
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 获取回复-带用户
  /v1/reply/update:
    post:
      consumes:
      - application/json
      description: 获取回复内容
      operationId: get-topic-content-json
      parameters:
      - description: 更新回复传入json
        in: body
        name: topic
        required: true
        schema:
          $ref: '#/definitions/service.UpdateReply'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 更新回复
  /v1/section/add:
    get:
      description: 添加板块
      operationId: get-section-add-json
      parameters:
      - description: 板块名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 添加板块
  /v1/section/all:
    get:
      description: 获取所有板块
      operationId: get-section-all-json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: 获取板块
  /v1/topic/add:
    post:
      consumes:
      - application/json
      description: 添加主题
      operationId: post-topic-add-json
      parameters:
      - description: 主题
        in: body
        name: topic
        required: true
        schema:
          $ref: '#/definitions/service.Topic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 添加主题
  /v1/topic/content/get:
    get:
      consumes:
      - application/json
      description: 获取主题内容
      operationId: get-topic-content-json
      parameters:
      - description: 主题ID
        in: query
        name: topicid
        required: true
        type: integer
      - description: 板块ID
        in: query
        name: sectionid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: 获取主题内容
  /v1/topic/delete:
    post:
      consumes:
      - application/json
      description: 删除主题
      operationId: get-topic-content-json
      parameters:
      - description: 更新主题传入json
        in: body
        name: topic
        required: true
        schema:
          $ref: '#/definitions/service.DeleteTopic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 删除主题
  /v1/topic/get:
    get:
      consumes:
      - application/json
      description: 根据页码等数据指定板块获取主题
      operationId: get-topic-page-json
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pagesize
        type: integer
      - description: 总数
        in: query
        name: total
        type: integer
      - description: 板块ID
        in: query
        name: sectionid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: 获取主题
  /v1/topic/get/{userid}:
    get:
      consumes:
      - application/json
      description: 根据页码等数据指定板块获取主题-带有用户权限
      operationId: get-topic-page-user-json
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pagesize
        type: integer
      - description: 总数
        in: query
        name: total
        type: integer
      - description: 板块ID
        in: query
        name: sectionid
        required: true
        type: integer
      - description: 用户ID
        in: path
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 获取主题-带用户
  /v1/topic/update:
    post:
      consumes:
      - application/json
      description: 获取主题内容
      operationId: get-topic-content-json
      parameters:
      - description: 更新主题传入json
        in: body
        name: topic
        required: true
        schema:
          $ref: '#/definitions/service.UpdateTopic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 更新主题
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"